print_function; move decals_wl to old
